Improve resilience against memory attacks
To reduce residual fragments of secret data in memory after
deallocation, this patch replaces the global delete operator with a
version that zeros out previously allocated memory. It makes use of
the new C++14 sized deallocation, but provides an unsized fallback
with platform-specific size deductions.

This change is only a minor mitigation and cannot protect against
buffer reallocations by the operating system or non-C++ libraries.
Thus, we still cannot guarantee all memory to be wiped after free.

As a further improvement, this patch uses libgcrypt to write
long-lived master key component hashes into a secure memory area
and wipe it afterwards.

Finally, it ensures that all password form fields are cleared when
they are not in use anymore and fixes a bug where removing a component
from the master key would not mark the database as dirty.